# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Intent Engine API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Lead {
  id: string;
  source: string;
  text: string;
  url?: string | null;
  contact?: string | null;
  intent_score: number;
  intent_label: string;
  subreddit?: string | null;
  author?: string | null;
  created_at: string;
  saved: boolean;
  status?: string | null;
}

export interface LeadsResponse {
  leads: Lead[];
  total: number;
  fetched_at: string;
}

export interface SourceCount {
  source: string;
  count: number;
}

export interface LeadsStats {
  total_leads: number;
  high_intent: number;
  medium_intent: number;
  low_intent: number;
  saved_count: number;
  by_source: SourceCount[];
  avg_score: number;
  last_run?: string | null;
}

export interface ReplyVariant {
  label: string;
  message: string;
  style: string;
}

export interface GeneratedResponse {
  message: string;
  lead_id: string;
  variants: ReplyVariant[];
}

export interface UpdateLeadStatusInput {
  status: string;
}

export interface UpdateLeadStatusResult {
  id: string;
  status: string;
}

export interface SaveLeadResult {
  saved: boolean;
  lead_id: string;
}

export interface SourceInfo {
  id: string;
  name: string;
  active: boolean;
  description: string;
  count: number;
}

export interface SourcesResponse {
  sources: SourceInfo[];
}

export interface Keyword {
  id: string;
  phrase: string;
  score: number;
  category: string;
  enabled: boolean;
  created_at: string;
}

export interface KeywordsResponse {
  keywords: Keyword[];
  total: number;
}

export interface CreateKeywordInput {
  phrase: string;
  score: number;
  category?: string;
}

export interface UpdateKeywordInput {
  phrase?: string;
  score?: number;
  category?: string;
  enabled?: boolean;
}

export interface TestPhraseInput {
  phrase: string;
}

export interface KeywordMatch {
  id: string;
  phrase: string;
  score: number;
}

export interface TestPhraseResult {
  phrase: string;
  score: number;
  intent_label: string;
  matched_keyword?: string | null;
  matched_keyword_id?: string | null;
  all_matches?: KeywordMatch[];
}

export interface DeleteKeywordResult {
  deleted: boolean;
  id: string;
}

export interface MatchedKeyword {
  phrase: string;
  score: number;
  /** True if this keyword set the final intent score (highest-scoring match) */
  is_primary: boolean;
}

export interface ScoreBreakdown {
  lead_id: string;
  final_score: number;
  matched: MatchedKeyword[];
  /** Number of active keywords that did NOT match this post */
  unmatched_count: number;
  total_keywords: number;
  /** True if no keywords matched and the default score of 3 was used */
  fallback: boolean;
}

export type LeadAnalysisPainLevel =
  (typeof LeadAnalysisPainLevel)[keyof typeof LeadAnalysisPainLevel];

export const LeadAnalysisPainLevel = {
  LOW: "LOW",
  MEDIUM: "MEDIUM",
  HIGH: "HIGH",
  CRITICAL: "CRITICAL",
} as const;

export type LeadAnalysisUrgency =
  (typeof LeadAnalysisUrgency)[keyof typeof LeadAnalysisUrgency];

export const LeadAnalysisUrgency = {
  LOW: "LOW",
  MEDIUM: "MEDIUM",
  HIGH: "HIGH",
} as const;

export type LeadAnalysisTechLevel =
  (typeof LeadAnalysisTechLevel)[keyof typeof LeadAnalysisTechLevel];

export const LeadAnalysisTechLevel = {
  NON_TECHNICAL: "NON_TECHNICAL",
  TECHNICAL: "TECHNICAL",
  DEVELOPER: "DEVELOPER",
} as const;

export type LeadAnalysisRecommendedChannel =
  (typeof LeadAnalysisRecommendedChannel)[keyof typeof LeadAnalysisRecommendedChannel];

export const LeadAnalysisRecommendedChannel = {
  POST_REPLY: "POST_REPLY",
  EMAIL: "EMAIL",
  DM: "DM",
} as const;

export type LeadAnalysisRecommendedStyle =
  (typeof LeadAnalysisRecommendedStyle)[keyof typeof LeadAnalysisRecommendedStyle];

export const LeadAnalysisRecommendedStyle = {
  EMPATHETIC: "EMPATHETIC",
  DIRECT: "DIRECT",
  CURIOSITY: "CURIOSITY",
} as const;

export interface LeadAnalysis {
  lead_id: string;
  summary: string;
  pain_level: LeadAnalysisPainLevel;
  urgency: LeadAnalysisUrgency;
  tech_level: LeadAnalysisTechLevel;
  recommended_channel: LeadAnalysisRecommendedChannel;
  recommended_style: LeadAnalysisRecommendedStyle;
  reasoning: string;
  key_angles: string[];
  avoid: string[];
  opening_hook: string;
}

export interface SourceProfile {
  karma: number | null;
  account_age_days: number | null;
  bio: string | null;
  website: string | null;
  profile_url: string | null;
}

export interface EnrichmentData {
  lead_id: string;
  emails: string[];
  phones: string[];
  urls: string[];
  company: string | null;
  source_profile: SourceProfile | null;
}

export interface VoiceConfig {
  id: string;
  businessName: string;
  businessType: string;
  greeting: string;
  instructions: string;
  hoursJson: string;
  servicesJson: string;
  voice: string;
  transferNumber?: string | null;
  twilioPhoneNumber?: string | null;
  twilioAccountSid?: string | null;
  twilioAuthToken?: string | null;
  isActive: boolean;
  webhookUrl: string;
  createdAt: string;
  updatedAt: string;
}

export interface UpdateVoiceConfigInput {
  businessName?: string;
  businessType?: string;
  greeting?: string;
  instructions?: string;
  hoursJson?: string;
  servicesJson?: string;
  voice?: string;
  transferNumber?: string | null;
  twilioAccountSid?: string | null;
  twilioAuthToken?: string | null;
  twilioPhoneNumber?: string | null;
  isActive?: boolean;
}

export interface VoiceCall {
  id: string;
  callSid: string;
  fromNumber: string;
  toNumber: string;
  direction: string;
  status: string;
  durationSeconds?: number | null;
  outcome?: string | null;
  summary?: string | null;
  startedAt: string;
  endedAt?: string | null;
  messageCount: number;
}

export interface VoiceCallsResponse {
  calls: VoiceCall[];
  total: number;
  page: number;
  totalPages: number;
}

export interface VoiceStatusCount {
  status: string;
  count: number;
}

export interface VoiceOutcomeCount {
  outcome: string;
  count: number;
}

export interface VoiceCallStats {
  totalCalls: number;
  todayCalls: number;
  avgDurationSeconds: number | null;
  inboundCount: number;
  outboundCount: number;
  byStatus: VoiceStatusCount[];
  byOutcome: VoiceOutcomeCount[];
}

export interface VoiceMessage {
  id: string;
  callId: string;
  role: string;
  content: string;
  audioReady: boolean;
  createdAt: string;
}

export interface VoiceCallDetail {
  call: VoiceCall;
  messages: VoiceMessage[];
}

export interface CreateOutboundCallInput {
  toNumber: string;
  purpose?: string | null;
}

export interface CreateOutboundCallResult {
  callSid: string;
  status: string;
  toNumber: string;
}

export type GetLeadsParams = {
  min_score?: number;
  source?: string;
  limit?: number;
  subreddit?: string;
};

export type GetVoiceCallsParams = {
  page?: number;
  limit?: number;
  direction?: string;
  status?: string;
};