# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Intent Engine API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /voice/config:
    get:
      operationId: getVoiceConfig
      tags: [voice]
      summary: Get voice agent business configuration
      responses:
        "200":
          description: Business configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceConfig"
    put:
      operationId: updateVoiceConfig
      tags: [voice]
      summary: Update voice agent business configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVoiceConfigInput"
      responses:
        "200":
          description: Updated configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceConfig"
...
  /voice/calls:
    get:
      operationId: getVoiceCalls
      tags: [voice]
      summary: Get paginated call history
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: direction
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Paginated call history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceCallsResponse"
  /voice/calls/{id}:
    get:
      operationId: getVoiceCall
      tags: [voice]
      summary: Get call detail with full transcript
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Call detail with transcript
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceCallDetail"
  /voice/outbound:
    post:
      operationId: createOutboundCall
      tags: [voice]
      summary: Initiate an outbound call
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOutboundCallInput"
      responses:
        "200":
          description: Outbound call initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateOutboundCallResult"
components:
  schemas:
...
    VoiceConfig:
      type: object
      properties:
        id:
          type: string
        businessName:
          type: string
        businessType:
          type: string
        greeting:
          type: string
        instructions:
          type: string
        hoursJson:
          type: string
        servicesJson:
          type: string
        voice:
          type: string
        transferNumber:
          type: string
          nullable: true
        twilioPhoneNumber:
          type: string
          nullable: true
        twilioAccountSid:
          type: string
          nullable: true
        twilioAuthToken:
          type: string
          nullable: true
        isActive:
          type: boolean
        webhookUrl:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - businessName
        - businessType
        - greeting
        - instructions
        - hoursJson
        - servicesJson
        - voice
        - isActive
        - webhookUrl
        - createdAt
        - updatedAt
    UpdateVoiceConfigInput:
      type: object
      properties:
        businessName:
          type: string
        businessType:
          type: string
        greeting:
          type: string
        instructions:
          type: string
        hoursJson:
          type: string
        servicesJson:
          type: string
        voice:
          type: string
        transferNumber:
          type: string
          nullable: true
        twilioAccountSid:
          type: string
          nullable: true
        twilioAuthToken:
          type: string
          nullable: true
        twilioPhoneNumber:
          type: string
          nullable: true
        isActive:
          type: boolean
    VoiceCall:
      type: object
      properties:
        id:
          type: string
        callSid:
          type: string
        fromNumber:
          type: string
        toNumber:
          type: string
        direction:
          type: string
        status:
          type: string
        durationSeconds:
          type: integer
          nullable: true
        outcome:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true
        startedAt:
          type: string
        endedAt:
          type: string
          nullable: true
        messageCount:
          type: integer
      required:
        - id
        - callSid
        - fromNumber
        - toNumber
        - direction
        - status
        - startedAt
        - messageCount
    VoiceCallsResponse:
      type: object
      properties:
        calls:
          type: array
          items:
            $ref: "#/components/schemas/VoiceCall"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - calls
        - total
        - page
        - totalPages
...
    VoiceMessage:
      type: object
      properties:
        id:
          type: string
        callId:
          type: string
        role:
          type: string
        content:
          type: string
        audioReady:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - callId
        - role
        - content
        - audioReady
        - createdAt
    VoiceCallDetail:
      type: object
      properties:
        call:
          $ref: "#/components/schemas/VoiceCall"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/VoiceMessage"
      required:
        - call
        - messages
    CreateOutboundCallInput:
      type: object
      properties:
        toNumber:
          type: string
        purpose:
          type: string
          nullable: true
      required:
        - toNumber
    CreateOutboundCallResult:
      type: object
      properties:
        callSid:
          type: string
        status:
          type: string
        toNumber:
          type: string
      required:
        - callSid
        - status
        - toNumber