# lib/api-zod/src/generated/types/createOutboundCallInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Intent Engine API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOutboundCallInput {
  toNumber: string;
  purpose?: string | null;
}