# lib/api-zod/src/generated/types/createOutboundCallResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Intent Engine API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOutboundCallResult {
  callSid: string;
  status: string;
  toNumber: string;
}