# lib/api-zod/src/generated/types/updateVoiceConfigInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Intent Engine API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateVoiceConfigInput {
  businessName?: string;
  businessType?: string;
  greeting?: string;
  instructions?: string;
  hoursJson?: string;
  servicesJson?: string;
  voice?: string;
  transferNumber?: string | null;
  twilioAccountSid?: string | null;
  twilioAuthToken?: string | null;
  twilioPhoneNumber?: string | null;
  isActive?: boolean;
}