# lib/api-zod/src/generated/types/voiceCall.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Intent Engine API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VoiceCall {
  id: string;
  callSid: string;
  fromNumber: string;
  toNumber: string;
  direction: string;
  status: string;
  durationSeconds?: number | null;
  outcome?: string | null;
  summary?: string | null;
  startedAt: string;
  endedAt?: string | null;
  messageCount: number;
}