# lib/api-zod/src/generated/types/voiceCallDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Intent Engine API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VoiceCall } from "./voiceCall";
import type { VoiceMessage } from "./voiceMessage";

export interface VoiceCallDetail {
  call: VoiceCall;
  messages: VoiceMessage[];
}