# lib/api-zod/src/generated/types/voiceCallsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Intent Engine API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VoiceCall } from "./voiceCall";

export interface VoiceCallsResponse {
  calls: VoiceCall[];
  total: number;
  page: number;
  totalPages: number;
}